+++ caqtus/gui/autogen/_editor_builder/__init__.py
@@ -1,4 +1,4 @@
-from ._attrs import build_attrs_class_editor
+from ._attrs import build_attrs_class_editor, AttributeOverride
 from ._editor_builder import EditorBuilder, TypeNotRegisteredError, EditorFactory
 
 __all__ = [
@@ -6,4 +6,5 @@ __all__ = [
     "TypeNotRegisteredError",
     "build_attrs_class_editor",
     "EditorFactory",
+    "AttributeOverride",
 ]