+++ caqtus/gui/autogen/_editor_builder/_attrs.py
@@ -1,19 +1,28 @@
-import inspect
-from typing import override
+from __future__ import annotations
+
+import functools
+import typing
+from collections.abc import Sequence
 
 import attrs
-import docstring_parser
 from PySide6.QtWidgets import QWidget, QFormLayout, QLabel
 
-from ._editor_builder import EditorBuilder, EditorBuildingError, EditorFactory
+from ._editor_builder import (
+    EditorBuilder,
+    EditorBuildingError,
+    EditorFactory,
+    TypeNotRegisteredError,
+)
+from .._int_editor import IntegerEditor
+from .._string_editor import StringEditor
 from .._value_editor import ValueEditor
 
 
 def build_attrs_class_editor[
     T: attrs.AttrsInstance
-](cls: type[T], builder: EditorBuilder, **attr_editors: EditorFactory) -> EditorFactory[
-    T
-]:
+](
+    cls: type[T], builder: EditorBuilder, **attr_overrides: AttributeOverride
+) -> EditorFactory[T]:
     """Build an editor for attrs class.
 
     This function will build a form editor with a list of widgets for the attributes of
@@ -22,82 +31,154 @@ def build_attrs_class_editor[
     The label for each widget is the name of the attribute, prettified by removing
     underscores and capitalizing the first letter of the first word.
 
-    The tooltip of this label is the docstring of the attribute.
-
     Args:
         cls: The attrs class to build the editor for.
         builder: The editor builder used to build editors for the class attributes.
-        **attr_editors: If a named argument corresponds to one of the attributes,
-            the editor passed for this argument will be used instead of using the
-            builder.
+        **attr_overrides: If a named argument corresponds to one of the
+            attributes, the object passed as argument will override the automated ui
+            generation for that attribute.
     """
 
-    fields: tuple[attrs.Attribute] = attrs.fields(cls)
+    fields: tuple[attrs.Attribute, ...] = attrs.fields(cls)
 
     if any(isinstance(field.type, str) for field in fields):
         # PEP 563 annotations - need to be resolved.
         attrs.resolve_types(cls)
 
-    attribute_docstrings = extract_attribute_docstrings(cls)
-
-    attribute_editors = {}
+    attr_ui_infos = []
     for field in fields:
-        if field.name in attr_editors:
-            attribute_editors[field.name] = attr_editors[field.name]
-            continue
-        if field.type is None:
-            raise AttributeEditorBuildingError(cls, field) from ValueError(
-                "No type specified"
-            )
+        override = attr_overrides.get(field.name, None)
         try:
-            attribute_editors[field.name] = builder.build_editor(field.type)
-        except EditorBuildingError as e:
+            label = get_attribute_label(field, override)
+            editor_factory = get_attribute_editor_factory(field, override, builder)
+            tooltip = get_attribute_tooltip(field, override)
+        except Exception as e:
             raise AttributeEditorBuildingError(cls, field) from e
+        attr_ui_infos.append(
+            AttributeUIInfo(
+                field_name=field.name,
+                label=label,
+                editor_factory=editor_factory,
+                tooltip=tooltip,
+            )
+        )
 
-    class AttrsEditor(ValueEditor[T]):
-        @override
-        def __init__(self) -> None:
-            self._widget = QWidget()
+    return functools.partial(AttrsEditor, cls, tuple(attr_ui_infos))
 
-            layout = QFormLayout()
-            self._widget.setLayout(layout)
-            for field in fields:
-                editor = attribute_editors[field.name]()
-                setattr(self, attr_to_editor_name(field.name), editor)
-                label = QLabel(prettify_snake_case(field.name))
-                if field.name in attribute_docstrings:
-                    label.setToolTip(attribute_docstrings[field.name])
-                layout.addRow(label, editor.widget())
 
-        @override
-        def set_value(self, value: T) -> None:
-            for field in fields:
-                editor = getattr(self, attr_to_editor_name(field.name))
-                assert isinstance(editor, ValueEditor)
-                editor.set_value(getattr(value, field.name))
+class AttrsEditor[T: attrs.AttrsInstance](ValueEditor[T]):
+    """A generic editor for attrs classes."""
 
-        # TODO: Figure out why pyright report this method as an incompatible override
-        @override
-        def read_value(self) -> T:  # type: ignore[reportIncompatibleMethodOverride]
-            attribute_values = {}
-            for field in fields:
-                editor = getattr(self, attr_to_editor_name(field.name))
-                assert isinstance(editor, ValueEditor)
-                attribute_values[field.name] = editor.read_value()
-            return cls(**attribute_values)
+    def __init__(self, cls: type[T], ui_specs: Sequence[AttributeUIInfo]) -> None:
+        self._widget = QWidget()
+        self._ui_specs = ui_specs
+        self._cls = cls
 
-        @override
-        def set_editable(self, editable: bool) -> None:
-            for field in fields:
-                editor = getattr(self, attr_to_editor_name(field.name))
-                assert isinstance(editor, ValueEditor)
-                editor.set_editable(editable)
+        layout = QFormLayout()
+        self._widget.setLayout(layout)
+        for ui_spec in ui_specs:
+            editor = ui_spec.editor_factory()
+            setattr(self, ui_spec.editor_name, editor)
+            label = QLabel(ui_spec.label)
+            if ui_spec.tooltip is not None:
+                label.setToolTip(ui_spec.tooltip)
+            layout.addRow(label, editor.widget())
 
-        @override
-        def widget(self) -> QWidget:
-            return self._widget
+    @typing.override
+    def set_value(self, value: T) -> None:
+        for ui_spec in self._ui_specs:
+            editor = getattr(self, ui_spec.editor_name)
+            assert isinstance(editor, ValueEditor)
+            editor.set_value(getattr(value, ui_spec.field_name))
 
-    return AttrsEditor
+    # TODO: Figure out why pyright report this method as an incompatible override
+    @typing.override
+    def read_value(self) -> T:  # type: ignore[reportIncompatibleMethodOverride]
+        attribute_values = {}
+        for ui_spec in self._ui_specs:
+            editor = getattr(self, ui_spec.editor_name)
+            assert isinstance(editor, ValueEditor)
+            attribute_values[ui_spec.field_name] = editor.read_value()
+        return self._cls(**attribute_values)
+
+    @typing.override
+    def set_editable(self, editable: bool) -> None:
+        for ui_spec in self._ui_specs:
+            editor = getattr(self, ui_spec.editor_name)
+            assert isinstance(editor, ValueEditor)
+            editor.set_editable(editable)
+
+    @typing.override
+    def widget(self) -> QWidget:
+        return self._widget
+
+
+@attrs.frozen
+class AttributeOverride:
+    label: str | None = None
+    editor_factory: EditorFactory | None = None
+    tooltip: str | None = None
+
+
+@attrs.frozen
+class AttributeUIInfo:
+    field_name: str
+    label: str
+    editor_factory: EditorFactory
+    tooltip: str | None
+
+    @property
+    def editor_name(self) -> str:
+        return attr_to_editor_name(self.field_name)
+
+
+def get_attribute_label(
+    attr: attrs.Attribute,
+    override: AttributeOverride | None,
+) -> str:
+    label = None
+    if override is not None:
+        label = override.label
+    if label is None:
+        label = prettify_snake_case(attr.name)
+    return label
+
+
+def get_attribute_editor_factory(
+    attr: attrs.Attribute,
+    override: AttributeOverride | None,
+    builder: EditorBuilder,
+) -> EditorFactory:
+    editor_factory = None
+    if override is not None:
+        editor_factory = override.editor_factory
+    if editor_factory is None:
+        if attr.type is None:
+            raise ValueError("Attribute has no type annotation")
+        try:
+            editor_factory = builder.build_editor(attr.type)
+        except TypeNotRegisteredError:
+            if attr.type is int:
+                editor_factory = IntegerEditor
+            elif attr.type is str:
+                editor_factory = StringEditor
+            else:
+                raise
+    return editor_factory
+
+
+def get_attribute_tooltip(
+    attr: attrs.Attribute,
+    override: AttributeOverride | None,
+) -> str | None:
+    tooltip = None
+    if override is not None:
+        tooltip = override.tooltip
+    # TODO: Extract tooltip from documentation
+    # Would need to parse the docstring of the class and find the attribute's docstring.
+    # In addition, might need to parse the docstring of parent classes if the attribute
+    # is inherited.
+    return tooltip
 
 
 def prettify_snake_case(name: str) -> str:
@@ -111,32 +192,6 @@ def prettify_snake_case(name: str) -> str:
 
 def attr_to_editor_name(name: str) -> str:
     return f"editor_{name}"
-
-
-def extract_attribute_docstrings(cls: type) -> dict[str, str]:
-    """Extract docstrings of attributes from a class.
-
-    Args:
-        cls: The class to extract the docstrings from.
-
-    Returns:
-        A dictionary mapping attribute names to their docstrings.
-
-        Not all attributes may have docstrings, so the dictionary may not contain all
-        attributes, or may be empty.
-    """
-
-    docstring = inspect.getdoc(cls)
-
-    if docstring:
-        parsed = docstring_parser.parse(docstring)
-        return {
-            param.arg_name: param.description
-            for param in parsed.params
-            if param.description is not None
-        }
-    else:
-        return {}
 
 
 class AttributeEditorBuildingError(EditorBuildingError):