+++ caqtus/types/units/_units.py
@@ -143,6 +143,7 @@ dimensionless = Dimensionless(BaseUnit(Unit("dimensionless")))
 TIME_UNITS = {"s", "ms", "µs", "us", "ns"}
 
 FREQUENCY_UNITS = {
+    "mHz",
     "Hz",
     "kHz",
     "MHz",
@@ -153,18 +154,21 @@ FREQUENCY_UNITS = {
 POWER_UNITS = {
     "W",
     "mW",
+    "µW",
+    "uW",
+    "nW",
     "dBm",
 }
 
-DIMENSIONLESS_UNITS = {"dB"}
+DIMENSIONLESS_UNITS = {"dB", "percent", "%"}
 
-CURRENT_UNITS = {"A", "mA"}
+CURRENT_UNITS = {"A", "mA", "µA", "uA", "nA"}
 
-VOLTAGE_UNITS = {"V", "mV"}
+VOLTAGE_UNITS = {"V", "mV", "µV", "uV", "nV"}
 
 DISTANCE_UNITS = {"m", "mm", "µm", "um", "nm"}
 
-DEGREE_UNITS = {"deg", "rad"}
+ANGLE_UNITS = {"deg", "°", "rad"}
 
 UNITS = (
     TIME_UNITS
@@ -174,7 +178,7 @@ UNITS = (
     | CURRENT_UNITS
     | VOLTAGE_UNITS
     | DISTANCE_UNITS
-    | DEGREE_UNITS
+    | ANGLE_UNITS
 )